+++ dali/operators/random/random_dist_test.cc
@@ -1,4 +1,4 @@
-// Copyright (c) 2025, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+// Copyright (c) 2025-2026, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -19,7 +19,7 @@
 #include <numeric>
 #include <random>
 #include <vector>
-#include "dali/operators/random/philox.h"
+#include "dali/core/random/philox.h"
 #include "dali/core/int_literals.h"
 #include "dali/core/span.h"
 #include "dali/operators/random/random_dist_test.h"
