+++ dali/operators/random/random_dist_test.cu
@@ -1,4 +1,4 @@
-// Copyright (c) 2025, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+// Copyright (c) 2025-2026, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -18,7 +18,7 @@
 #include <vector>
 #include "dali/core/cuda_stream.h"
 #include "dali/core/dev_buffer.h"
-#include "dali/operators/random/philox.h"
+#include "dali/core/random/philox.h"
 #include "dali/operators/random/random_dist_test.h"
 
 namespace dali {
